FileProvider ssh client extension for proxy connections
- Made Code and FileProvider work with the same proxy dial that will
automatically start a thread and take care of the connection.
- Small refactor on Code.
