+++ BlinkCode/CodeFileSystem.swift
@@ -29,201 +29,12 @@
 //
 ////////////////////////////////////////////////////////////////////////////////
 
-
 import Combine
 import Foundation
 import Network
 
 import BlinkFiles
-import BlinkConfig
-import System
-import SwiftUI
 
-struct MountEntry: Codable {
-  let name: String
-  let root: String
-}
-
-class TranslatorReference {
-  let translator: Translator
-  let cancel: () -> Void
-
-  init(_ translator: Translator, cancel: @escaping (() -> Void)) {
-    self.translator = translator
-    self.cancel = cancel
-  }
-  
-  deinit {
-    cancel()
-  }
-}
-
-public class CodeFileSystemService: CodeSocketDelegate {
-  
-  let server: WebSocketServer
-  let log: BlinkLogger
-  
-  public let port: UInt16
-  var tokens: [Int: MountEntry] = [:]
-  var tokenIdx = 0;
-
-  private var translators: [String: TranslatorReference] = [:]
-
-  private let finishedCallback: ((Error?) -> ())
-  func finished(_ error: Error?) { finishedCallback(error) }
-  
-  public var state: NWListener.State {
-    server.listener.state
-  }
-  
-  public func registerMount(name: String, root: String) -> Int {
-    tokenIdx += 1
-    tokens[tokenIdx] = MountEntry(name: name, root: root)
-    log.info("Registered mount \(tokenIdx) for \(name) at \(root)")
-    return tokenIdx
-  }
-
-  public func deregisterMount(_ token: Int) {
-    log.info("De-registering mount \(tokenIdx)")
-    // If no other token is using the same translator, trash it
-    guard let token = tokens.removeValue(forKey: tokenIdx) else {
-      return
-    }
-    let root = URL(string: token.root)!
-
-    // If we have no host, there is no remote translator
-    guard let host = root.host,
-          let _ = translators[host] else {
-      return
-    }
-
-    // Remove the Translator if no other mounts use it.
-    if let _ = tokens.first(where: { (_, tk) in
-                                   let url = URL(string: tk.root)!
-                                   return host == url.host
-                                 }) {
-      return
-    }
-
-    translators.removeValue(forKey: host)
-  }
-
-  public init(listenOn port: NWEndpoint.Port, tls: Bool, finished: @escaping ((Error?) -> ()))  throws {
-    self.port = port.rawValue
-    self.server = try WebSocketServer(listenOn: port, tls: tls)
-    self.finishedCallback = finished
-    
-    self.log = CodeFileSystemLogger.log("FileSystem")
-    
-    self.server.delegate = self
-  }
-  
-  func getRoot(token: Int, version: Int) -> WebSocketServer.ResponsePublisher {
-    if let mount = self.tokens[token] {
-      return .just((try! JSONEncoder().encode(mount), nil)).eraseToAnyPublisher()
-    } else {
-      return .just((nil, nil)).eraseToAnyPublisher()
-    }
-  }
-
-  public func handleMessage(encodedData: Data, binaryData: Data?) -> WebSocketServer.ResponsePublisher {
-    guard let request = try? JSONDecoder().decode(BaseFileSystemRequest.self, from: encodedData) else {
-      log.error(String(data: encodedData, encoding: .utf8) ?? "Error decoding data")
-      return .fail(error: WebSocketError(message: "Bad request"))
-    }
-
-    do {
-      switch request.op {
-      case .getRoot:
-        let msg: GetRootRequest = try decode(encodedData)
-        return try getRoot(token: msg.token, version: msg.version)
-      case .stat:
-        let msg: StatFileSystemRequest = try decode(encodedData)
-        return try fileSystem(for: msg.uri).stat()
-      case .readDirectory:
-        let msg: ReadDirectoryFileSystemRequest = try decode(encodedData)
-        return try fileSystem(for: msg.uri).readDirectory()
-      case .readFile:
-        let msg: ReadFileFileSystemRequest = try decode(encodedData)
-        return try fileSystem(for: msg.uri).readFile()
-      case .writeFile:
-        let msg: WriteFileSystemRequest = try decode(encodedData)
-        return try fileSystem(for: msg.uri).writeFile(options: msg.options,
-                                                      content: binaryData ?? Data())
-      case .createDirectory:
-        let msg: CreateDirectoryFileSystemRequest = try decode(encodedData)
-        return try fileSystem(for: msg.uri).createDirectory()
-      case .rename:
-        let msg: RenameFileSystemRequest = try decode(encodedData)
-        return try fileSystem(for: msg.oldUri).rename(newUri: msg.newUri,
-                                                     options: msg.options)
-      case .delete:
-        let msg: DeleteFileSystemRequest = try decode(encodedData)
-        return try fileSystem(for: msg.uri).delete(options: msg.options)
-      }
-    } catch {
-      log.error("\(error)")
-      log.error("Processing request \(String(data: encodedData, encoding: .utf8) ?? "Error decoding data")")
-      return .fail(error: error)
-    }
-  }
-
-  private func fileSystem(for uri: URI) throws -> CodeFileSystem {
-    let rootPath = uri.rootPath
-
-    if let host = rootPath.host,
-       let tRef = translators[host],
-       tRef.translator.isConnected {
-      return CodeFileSystem(.just(tRef.translator), uri: uri)
-    }
-    
-    switch(rootPath.protocolIdentifier) {
-    case "blinksftp":
-      let builder = TranslatorFactories.sftp
-      var thread: Thread!
-      var runLoop: RunLoop? = nil
-      
-      let threadIsReady = Future<RunLoop, Error> { promise in
-        thread = Thread {
-          let timer = Timer(timeInterval: TimeInterval(1), repeats: true) { _ in
-            //print("timer")
-          }
-          runLoop = RunLoop.current
-          RunLoop.current.add(timer, forMode: .default)
-          promise(.success(RunLoop.current))
-          CFRunLoopRun()
-          // Wrap it up
-          RunLoop.current.run(until: Date(timeIntervalSinceNow: 0.5))
-        }
-        thread.start()
-      }
-      
-      guard let hostAlias = rootPath.host else {
-        throw WebSocketError(message: "Missing host on rootpath")
-      }
-
-      let translator = threadIsReady
-        .flatMap { builder.buildOn($0, hostAlias: hostAlias) }
-        .map { t -> Translator in
-          self.translators[rootPath.host!] = TranslatorReference(t, cancel: {
-            self.log.debug("Cancelling translator")
-            let cfRunLoop = runLoop!.getCFRunLoop()
-            CFRunLoopStop(cfRunLoop)
-          })
-          return t
-        }
-        .eraseToAnyPublisher()
-
-      return CodeFileSystem(translator, uri: uri)
-
-    case "blinkfs":
-      // The local one does not need to be saved.
-      return CodeFileSystem(TranslatorFactories.local.build(rootPath), uri: uri)
-    default:
-      throw WebSocketError(message: "Unknown protocol - \(rootPath.protocolIdentifier)")
-    }        
-  }  
-}
 
 class CodeFileSystem {
   private let translator: AnyPublisher<Translator, Error>
@@ -492,59 +303,4 @@ class CodeFileSystem {
       .eraseToAnyPublisher()
   }
 
-}
-
-extension CodeFileSystemService {
-  func decode<T: Decodable>(_ encodedData: Data) throws -> T {
-    try JSONDecoder().decode(T.self, from: encodedData)
-  }
-}
-
-extension RootPath {
-  public var parent: RootPath {
-    return RootPath(url.deletingLastPathComponent())
-//    RootPath([protocolIdentifier,
-//              host,
-//              (filesAtPath as NSString).deletingLastPathComponent]
-//              .compactMap { $0 }
-//              .joined(separator: ":"))
-  }
-}
-
-struct CodeFileSystemLogger {
-  static var handler = [BlinkLogging.LogHandlerFactory]()
-  static func log(_ component: String) -> BlinkLogger {
-    if handler.isEmpty {
-      handler.append(
-        {
-          $0.format { [ $0[.component] as? String ?? "global",
-                      $0[.message] as? String ?? ""
-                    ].joined(separator: " ") }
-          .sinkToOutput()
-        }
-      )
-      
-      let dateFormatter = DateFormatter()
-      dateFormatter.dateFormat = "MMM dd YYYY, HH:mm:ss"
-      if let file = try? FileLogging(to: BlinkPaths.blinkCodeErrorLogURL()) {
-        handler.append(
-          {
-            try $0
-            .filter(logLevel: .debug)
-            .format {
-              [ "[\($0[.logLevel]!)]",
-                dateFormatter.string(from: Date()),
-                $0[.component] as? String ?? "global",
-                $0[.message] as? String ?? ""
-              ].joined(separator: " ") }
-            .sinkToFile(file)
-          }
-        )
-      } else {
-        print("File logging not working")
-      }
-    }
-
-    return BlinkLogger(component, handlers: handler)
-  }
 }
